testing if selfish sending is similar to optimizing fee potential globally
some experiments to understand if the global optimization of the fee potential makes any sense. We observe that the the selfish behavior of sending nodes quickly produces a network with depleted channels. those seem to be located similarily to those when the global optimization problem is solved. This is evidence that it makes sense to study this optimization problem. also on the notebook it is visible that selfish behavior and the depleted channels lowers the max flow statistic overall